+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 ## NAACCR XML Version History
 
+**Version 4.11**
+
+- Updated commons-lang3 dependency from version 3.6 to version 3.7.
+- Updated commons-io dependency from version 2.5 to version 2.6.
+- Updated XZ compression library from version 1.6 to version 1.8.
+
 **Version 4.10**
 
 - Added maximum length of 50 characters for NaaccrId attribute in the dictionaries.